Basically done
Will be committing a lot just in case today. Ran into a huge scare, but the project is pretty good, and I am happy with it. It runs smooth and look decent plus the AI works as attended. Will try to add sound, some tests and comments by the end of today.